ignoreファイルを移動
