+++ .gitignore
@@ -1,3 +1,4 @@
 test.rb
 .ruby-version
 *.txt
+archive
